Safari-609.1.11.1/LayoutTests/fast/events/touch/ios/no-focus-change-when-preventing-default-on-touchstart-expected.txt

Verifies that focus doesn't move into the subframe when preventDefault() is invoked on touchstart. To test manually, tap the button and check that the top level frame does not lose focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.activeElement is document.body
PASS successfullyParsed is true

TEST COMPLETE
